adding sourceLoc support for "Match" command
you can now specify sourceLoc: x1,y1, w, h as part of the match command
(in the same level as "match: filename.png". If specified, templateMatch
will ONLY search in that location (to make it work, the region is 1 pix
wider and taller than the signature).

This takes about 0.7 sec per match, rather than 1.x second. Not a large
improvement.

Possibility to speed things up: to extract the sig only once during the
reading process, instead of doing it for each match.

Will fix in the next version.