fixed the problem with skimming & gaze exporting
With the issue that vt (for video time) reporting is incorrect -- it
reflects the video time of processing. So if I skimmed forward 500ms and
there wasn't any screen change, I will process the gaze data in the past
500ms, and the VT will be the time of processing -- i.e., I can't tell
from vt when the fixation happened.

I can fix that by back-tracking -- using gazetime-offset to proximate
this. This will be implemented in the next version.