Eliminated tracking; use match instead
No need to keep match and track separately. They are essentially the
same thing.

As implemented now, when sourceLoc is specified, we do a comparison at
the same location. When it's not specified, we do a full screen search,
i.e., track.

Added/expanded support for relativeAOI. Now matched sig will generate a
special AOI with id="__MATCH__". When you specify a "relativeAOI", the
xywh are relative to the top-left corner of the matched sig, instead of
a regular AOI that uses absolute coordinates.

The track code is still kept in the source; will remote later.

To-do: the logic should be:
Match: source.png
sourceLoc: xyhw
destRange: xyhw, which can be identical to above or a different range.
We then search in that range. If missing, the default is ... either
sourceLoc or full screen.