Major change: eyelink.awk uses a different time estimate
The time stamp in the eyelink *_eye.log files begins after the
calibration; but the video actually includes the calibration, which can
vary from 30 sec to 5 min. The result is total async in the gaze-video.

The solution is to use the first column in the _eye.log, which is the
time of log entry. It tracks the calibration time, etc. When we use this
as the time stamp, without any further adjustments, we can get close to
1sec or less.