adding the last AOI ID to the "page" field
The item right below Assessment/items used to show up on in the AOIid
field, but not the "page" or context field. In this case when the gaze
is NOT on a particular AOI, the "page" field says only Assessment/Items,
not showing which page it's on. This change will address that problem.
It also introduce the issue when you have nested AOIs uner AOIs, in
which case this field is no longer "page". We have to deal with that by
redefining "page" as the stuff after Assessment/Items.