various fixes
Fixing a few minor things, but also trying to rethink how to refactor
the p2YAML. In theory I shouldn't do yamlconfig[] processing repeated in
the p2Task(); I can move it to the p2ReadSigImage() to run once.

On the other hand, when I do line_profile, the time saved is not worth
it.