[dirty] Adding videoScalingFactor functions and AOI fixes
- Cleaned up getMousePositionsFromVideo()
- Implemented getVideoScalingFactors ()
- Changed getCurrentAOIs() to add the lastVTime constraint
- Changed the signatureImageDict structure, to allow parameters about a
sig to be stored
- Implemented an algorithm to first search at last known location, and
search the whole place if not found. **unfortunately, this won't payoff
unless very late in the video when most sigs have been identified. I can
hack this for ITDS because the sigs for all pages are at the identical
location.
- also tried to refactor the "match" part of the p2Task(); this produces
some weird output that I don't have time to debug. Move it back for now.

- enhancing various docstrings to add :param etc.