[dirty] adding textMatch() and a bunch of other fixes
- changed import to avoid from X import *; as a result we introduce a
couple of module names

- removed doTemplateMatching() because it's duplicated from the 'match'
in value: code and won't run correctly.

- FOR ITDS only: in matching Assessment/items, once one match is found,
all other items are set to the same lastKnownPosition. Not going to work
for other projects!!

- For Proportional Punch, I had to create 'textMatch' to use ocr of item
numbers as the item id

- added YAML code for topleftTemplate and bottomrightTemplate parameters
for AOI shift/scaling factors. Note this is problematic if not all AOIs
are scaled to the same factor. Ideally we need to move to a per-AOI
setting.