a number of unfinished fixes
- mousevideodata: returns none if no enough valid data in the sample

- getVideoScalingFactors: adding threshold parameter and adding/showing
the AOIs

- @@Hack: Trying to deal with the case when the whole CBAL window is
moved on the screen. Logically we need to reestimate the
scaling/shifting parameters per frame, which will be a chore. For now, I
am just treating the lastMatchOffset as the new aoiShiftX/Y. This is not
a very good idea.

- getVideoScalingFactors: adding windowName as a parameter for
displaying; also fixing a bug in mvdStartTime, which for empty input can
give errors.

- logging: using removeHandler() rather than shutdown().