fixing a bug in output:
We have added subjnum and vTime in the aoi list;
in logEvents(), they are printed out as the first 2 fields in the
aoistring. But in the cases of no-aoi-match, the format is different --
starting with JUNKSCREEN or soemthing like that rather than subjnum.

Fixed by not outputing subject and vtime.